Fix footer in Chromium-based browsers
On Chromium-based browsers, antialiasing is ugly, so clipping to the bottom of the element leaves bad antialiasing and a visible line at the edge where the drop shadow would start on the bottom edge. Adding a few pixels to the height, and clipping those back off with clip-path completely eliminates the bottom border causing this. This issue doesn't appear in Firefox, but does in Chromium-based browsers.